Compute the chorus parameters in double, as upstream does
Every field the chorus setters write is already a double, but each one
was computed in float first and then widened -- so the stored value was
float32's answer.  The LFO rate is the one that shows: rateInc is
accumulated into the phase once per sample, so a float rate drifts the
phase further the faster it runs, and gs_chorus_rate sweeps the rate
across its whole range over 143 seconds.  The error grew monotonically
to -78 dBFS by the end.

The pre-lowpass also used powf and expf where upstream uses Math.pow and
Math.exp, and sampleRate was stored as a float.

gs_chorus_rate 67.05 -> 131.96 dB, gs_chorus_level 149.01, STORM 104.14
-> 105.68. The corpus is now 2 very-close and 29 near-identical.

+++ spessasynth_core/include/spessasynth/synthesizer/dsp/chorus.h
@@ -37,7 +37,7 @@ typedef struct {
 	double preLPFz;
 	float *leftDelayBuffer;
 	float *rightDelayBuffer;
-	float sampleRate;
+	double sampleRate;
 	double phase;
 	unsigned int write;
 	double gain;
